backend-repo/requirements.txt
fastapi
uvicorn
faster-whisper
transformers
torch
ffmpeg-python
pytesseract
python-docx
fpdf2
sqlalchemy
pydantic
python-dotenv
python-multipart
yt-dlp
numpy
pillow
opencv-python
groq
google-generativeai
pymysql
psycopg2-binary
psutil
